Move getItemTypeAndMetaLabel to mixin
