+++ bundles/org.openhab.ui/web/src/pages/settings/items/items-list-vlist.vue
@@ -132,7 +132,10 @@
 </style>
 
 <script>
+import ItemMixin from '@/components/item/item-mixin'
+
 export default {
+  mixins: [ItemMixin],
   components: {
     'empty-state-placeholder': () => import('@/components/empty-state-placeholder.vue')
   },
@@ -250,24 +253,6 @@ export default {
         }
       }
       return tagsNonS
-    },
-    getItemTypeAndMetaLabel (item) {
-      let ret = item.type
-      if (item.metadata && item.metadata.semantics) {
-        ret += ' · '
-        const classParts = item.metadata.semantics.value.split('_')
-        ret += classParts[0]
-        if (classParts.length > 1) {
-          ret += ' > ' + classParts.pop()
-          if (item.metadata.semantics.config && item.metadata.semantics.config.relatesTo) {
-            const relatesToParts = item.metadata.semantics.config.relatesTo.split('_')
-            if (relatesToParts.length > 1) {
-              ret += ' · ' + relatesToParts.pop()
-            }
-          }
-        }
-      }
-      return ret
     },
     toggleCheck () {
       this.showCheckboxes = !this.showCheckboxes
